+++ plugin/commands/index.js
@@ -1,6 +1,7 @@
 exports.ping = require('./ping');
 exports.waypoint = require('./waypoint');
 exports.boatinfo = require('./boatinfo');
+exports.ask = require('./ask');
 
 exports.help = {
   example: 'Help',